+++ app/components/bali/hover_card/preview.rb
@@ -3,39 +3,86 @@
 module Bali
   module HoverCard
     class Preview < ApplicationViewComponentPreview
-      def with_hover_url(hover_url: '/show-content-in-hovercard')
-        render HoverCard::Component.new(hover_url: hover_url) do |c|
+      # Default
+      # --------
+      #
+      # @param placement select [auto, auto-start, auto-end, top, top-start, top-end, bottom, bottom-start, bottom-end, right, right-start, right-end, left, left-start, left-end]
+      # @param open_on_click toggle
+      # @param content_padding toggle
+      def default(placement: 'auto', open_on_click: false, content_padding: true)
+        render HoverCard::Component.new(
+          placement: placement,
+          open_on_click: open_on_click,
+          content_padding: content_padding
+        ) do |c|
           c.trigger do
             tag.p('Hover me!', class: 'has-text-centered')
           end
+
+          c.tag.p('Hovercard content!')
         end
       end
 
-      def with_template
-        render HoverCard::Component.new do |c|
+      # @!group Placements
+      def top
+        render HoverCard::Component.new(placement: 'top') do |c|
           c.trigger do
             tag.p('Hover me!', class: 'has-text-centered')
           end
 
-          c.template do
-            tag.p('Hovercard content!')
-          end
+          c.tag.p('Hovercard content!')
         end
       end
 
-      # Hover Card with option open-on-click
+      def right
+        render HoverCard::Component.new(placement: 'right') do |c|
+          c.trigger do
+            tag.p('Hover me!', class: 'has-text-centered')
+          end
+
+          c.tag.p('Hovercard content!')
+        end
+      end
+
+      def bottom
+        render HoverCard::Component.new(placement: 'bottom') do |c|
+          c.trigger do
+            tag.p('Hover me!', class: 'has-text-centered')
+          end
+
+          c.tag.p('Hovercard content!')
+        end
+      end
+
+      def left
+        render HoverCard::Component.new(placement: 'left') do |c|
+          c.trigger do
+            tag.p('Hover me!', class: 'has-text-centered')
+          end
+
+          c.tag.p('Hovercard content!')
+        end
+      end
+
+      # @!endgroup
+
+      # With Hover URL
       # --------------
-      # This option will render the HoverCard component but to show you the content, you have to
-      # click the link
-      # @param open_on_click [Boolean]
-      def open_on_click(open_on_click: true)
-        render HoverCard::Component.new(open_on_click: open_on_click) do |c|
+      # Content is loading asyncronously from the provided url
+      #
+      # @param placement select [auto, auto-start, auto-end, top, top-start, top-end, bottom, bottom-start, bottom-end, right, right-start, right-end, left, left-start, left-end]
+      # @param open_on_click toggle
+      # @param content_padding toggle
+      def with_hover_url(hover_url: '/show-content-in-hovercard', placement: 'auto',
+                         open_on_click: false, content_padding: true)
+        render HoverCard::Component.new(
+          hover_url: hover_url,
+          placement: placement,
+          open_on_click: open_on_click,
+          content_padding: content_padding
+        ) do |c|
           c.trigger do
-            tag.p('Click me!', class: 'has-text-centered')
-          end
-
-          c.template do
-            tag.p('Hovercard content!')
+            tag.p('Hover me!', class: 'has-text-centered')
           end
         end
       end