+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.40.1] - 2022-08-18
+
+Add `GanttChart::TaskActions::Component` displaying a menu with options for each task:
+
+- Opening details
+- Indent
+- Outdent
+- Delete
+
+Refactor `HoverCard::Component` to use tippy to simplify and handle more options.
+
 ## [0.40.0] - 2022-08-18
 
 Create `GanttChart::Component` for a full fledged Gantt Chart with the following functionality: